split profiles.py into separate sections by topic.
